Add files via upload
ok so this is what i did, but for the creation of the tree later on im pretty sure i fked up bcz my computer refused to run the code, like it remains as [*] forever so i'm gna sleep and let it rest and come back. hopefully my computer will let me see the heatmap and stuff so i can ownself see what went wrong....like personally i cannot visualize well so i have to see it and i can't. i'm so sorry for the inconvenience